+++ projectroles/tests/test_app_links.py
@@ -494,7 +494,7 @@ class TestAppLinkAPI(ProjectMixin, RoleAssignmentMixin, ViewTestBase):
             },
             {
                 'name': 'admin',
-                'url': '/admin/',
+                'url': None,  # No URL for Django admin, opens warning modal
                 'label': 'Django Admin',
                 'icon': 'mdi:cogs',
                 'active': False,