+++ CHANGELOG.rst
@@ -110,6 +110,7 @@ Fixed
     - Invalid ``role`` values in ``ProjectDetailView`` context data (#1706)
     - Incorrect icon placement in DataTables custom pagination links (#1716)
     - Plugin ``search()`` order not respected by results view DataTables (#1699, #1720)
+    - Django admin warning modal not displayed (#1730)
 
 Removed
 -------